Yaml input support

+++ test/runTests.test.js
@@ -5,15 +5,29 @@ const fs = require("fs");
 const artifactPath = path.resolve(__dirname, "./artifacts");
 const outputFile = path.resolve(`${artifactPath}/testResults.json`);
 
-describe("Run tests sucessfully", function () {
+describe("Run tests successfully", function () {
   // Set indefinite timeout
   this.timeout(0);
-  it("All specs pass", async () => {
-    await spawnCommand(
-      `node ./src/index.js runTests -c ${artifactPath}/config.json -i ${artifactPath} -o ${outputFile}`
-    );
-    const result = require(outputFile);
-    fs.unlinkSync(outputFile);
-    assert.equal(result.summary.specs.fail, 0);
+  const configFiles = [
+    { name: "JSON", file: "config.json" },
+    { name: "YAML", file: "config.yaml" }
+  ];
+
+  configFiles.forEach(({ name, file }) => {
+    it(`All specs pass with ${name} config`, async () => {
+      const configPath = path.join(artifactPath, file);
+      
+      // Ensure the config file exists
+      if (!fs.existsSync(configPath)) {
+        throw new Error(`Config file not found: ${configPath}`);
+      }
+
+      await spawnCommand(
+        `node ./src/index.js runTests -c ${artifactPath}/config.json -i ${artifactPath} -o ${outputFile}`
+      );
+      const result = require(outputFile);
+      fs.unlinkSync(outputFile);
+      assert.equal(result.summary.specs.fail, 0);
+    });
   });
 });
